src/requirements.txt
pystray>=0.19
Pillow>=10.0
pywinpty>=2.0
